finished adding comments to my components and separated the comments on Route view into its own component
