+++ src/components/comments/comments.css
@@ -1,0 +1,82 @@
+.comment-container{
+    max-height: 210px;
+    overflow: scroll;
+}
+.comment-box{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    width: 98%;
+    padding: 7px;
+    margin:3px;
+    border: .5px solid rgb(196, 193, 193);
+    border-radius: 5px;
+}
+.comment-image-body{
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+}
+.comment-user-image{
+    border: 1px solid #333640;
+    border-radius: 15px;
+    height: 30px;
+    width: 30px;
+    margin-right: 5px;
+    margin-top: 3px;
+    align-self: flex-start;
+    background-size: contain;
+    background-repeat: no-repeat;
+    background-position: center;
+}
+
+.comment-owner-name{
+    font-size: .8em;
+    margin:0px;
+    margin-top: 5px;
+    font-weight: bold;
+    color: #333640;
+}
+.comment-created-at{
+    color: grey;
+    font-size: .5em;
+    margin:0px;
+}
+.comment-body{
+    font-size: .8em;
+    margin:0px;
+    margin-top: 3px; 
+    color: #333640; 
+}
+.new-comment-body{
+    border: .5px solid rgb(196, 193, 193);
+    border-radius: 5px;
+    width: 98%;
+    height: 40px;
+    margin:3px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+}
+.new-comment-body input{
+    height: 38px;
+    width: 85%;
+    margin-left: 2px;
+    border: none;
+    padding-left: 5px;
+}
+.post-comment-button{
+    height: 25px;
+    border: .5px solid rgb(196, 193, 193);
+    border-radius: 5px;
+    margin-right: 10px;
+    background-color: rgb(230, 230, 231);
+    padding: 5px;
+}
+.remove-comment-button{
+    align-self: flex-start;
+    background: none;
+    border: none;
+}
