+++ src/components/routeViewButtons/RouteViewButtons.js
@@ -12,6 +12,8 @@ class RouteViewButtons extends Component {
       isTodo: false
     }
   }
+  //ticks are on redux so I run through those to see if the route being views is on the ticked list
+  //same with todos
   componentDidMount(){
     this.props.ticks.forEach(tick => {
       if(tick.route_id === this.props.route_id){
@@ -27,27 +29,33 @@ class RouteViewButtons extends Component {
         })
     })
   }
+  //remove tick through redux and set state.... should I use promise.resolve here?
   removeTick(){
     this.props.removeTick(this.props.route_id)
     this.setState({
       ticked: false,
     })
   }
+  //this is opens the add tickpopup which controls the addTick functionality.
+  // if someone is ticking a route... it should no longer be a todo so I remove the todo from state
+  //I feel like I should handle this better.... maybe re render the component to check the new todo's list instead of just change 
+  //local state right away. what if they exit out of the modal...
   tickButton(){
-    this.props.removeTodo(this.props.route_id)
     this.setState({
       ticked: true,
       isTodo: false
     })
     this.props.AddTickVisibility('visible')
   }
+
+  //send addTodo through redux... use promise.resolve?
   addTodo(){
-    console.log(this.props.gyms.selectedGym)
     this.props.addTodo(this.props.route_id, this.props.gyms.selectedGym)
     this.setState({
       isTodo: true,
     })
   }
+  //remove todo through redux... user promise.resolve?
   removeTodo(){
     this.props.removeTodo(this.props.route_id)
     this.setState({
