+++ src/views/dashboard/dashboard.css
@@ -46,6 +46,3 @@
     padding: 7px;
     width: 200px;
 }
-.type-selector{
-   
-}
