+++ src/views/profile/Profile.js
@@ -2,6 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import Header from '../../components/header/Header';
 import Gyms from '../../components/gyms/Gyms';
+import columns from '../../helperFiles/tableColumns';
 import ReactTable from "react-table";
 import "react-table/react-table.css";
 import './profile.css'
@@ -12,44 +13,6 @@ class Profile extends Component {
     render() {
       let ticks = this.props.ticks.filter(tick=>tick.gym_id === this.props.gyms.selectedGym)
       let todos = this.props.todos.filter(todo=>todo.gym_id === this.props.gyms.selectedGym)
-      let columns = [
-        {
-            Header:'',
-            id: 'color',
-            accessor: row => <div className='color-box' style={{backgroundColor: row.color}} />,
-            minWidth: 29,
-            maxWidth: 29
-        },
-        {
-            Header: 'Grade',
-            accessor: `difficulty`,
-            minWidth: 53,
-            maxWidth: 53
-         },
-         {
-             Header: 'Wall',
-             accessor: 'wall',
-             minWidth: 63,
-             maxWidth: 80
-         },
-         {
-             Header: 'Setter',
-             accessor: 'setters_name',
-             minWidth: 70,
-             maxWidth: 95
-         },
-         {
-             Header: 'Stars',
-             id:'avg_stars',
-             accessor: row => { let stars =[]
-                                 for(let i=0;i<row.avg_stars;i++){stars.push(i)}
-                                return <div className='stars-container'> {stars.map((star,i)=><div key={i} className='star-icon'/>)}</div>
-                             },
-             minWidth: 50,
-             maxWidth: 60
-         }
-    ]
-
       return (
         <div>
             <Header />
@@ -116,9 +79,6 @@ class Profile extends Component {
 
                 </div>
                 }
-                {/* <div className='popup'>
-                    <AddGym visibility={this.state.addGymPopup} AddGymVisibility={(value)=>this.AddGymVisibility(value)}/>
-                </div> */}
             </div>
         </div>
       );
