+++ src/views/profile/profile.css
@@ -18,7 +18,6 @@
     font-size: 1.3em;
     color: #3c3e47;
     letter-spacing: .03em
-    
 }
 .arrow-div{
     display: flex;
